requirements.txt
numpy==1.20.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
Pillow==8.4.0
PyYAML==6.0
regex==2022.9.13